python-utils/pyproject.toml
[tool.poetry]
name = "vcutils"
version = "0.7.2"
description = "Python general utility package for VCell"
authors = ["Virtual Cell"]
license = "MIT"
repository = "https://github.com/virtualcell/vcell"
keywords = ["HDF5", "SBML", "VCell"]
classifiers = [
    'Development Status :: 3 - Alpha',
    'Intended Audience :: Science/Research',
    'License :: OSI Approved :: MIT License',
    'Topic :: Scientific/Engineering :: Bio-Informatics',
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3 :: Only",
]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.31.0"
pydantic = "^2.5.2"
python-dotenv = "^1.0.0"
typer = "^0.9.0"
h5py = "^3.10.0"
rich = "^13.7.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
mypy = "^1.4.1"

[tool.poetry.group.dev.dependencies]
types-requests = "^2.31.0.10"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
